refactor(runbotics-orchestrator-ui): :art: code refactor after CR
